feat: enable KVBM emit metrics in Dynamo TRTLLM
